Refactor the user-button structure
