+++ packages/ui/src/mosaic/user-button/user-button.controller.tsx
@@ -1,307 +1,206 @@
-import { getFullName, getIdentifier } from '@clerk/shared/internal/clerk-js/user';
-import { useClerk, useOrganization, usePortalRoot, useSession, useUser } from '@clerk/shared/react';
-import type { CustomPage, OrganizationResource, UserResource } from '@clerk/shared/types';
+import { useSpinDelay } from '../hooks/useSpinDelay';
+import { setup } from '../machine/setup';
+import { useMachine } from '../machine/useMachine';
+import type { UserButtonModel } from './user-button.model';
+import type { UserButtonMenuProps, UserButtonModeProps } from './user-button.types';
+import type { UserButtonProps as UserButtonViewProps, UserButtonTriggerProps } from './user-button.view';
+import { userButtonBusyKeys } from './user-button.view';
 
-import { populateParamFromObject } from '../../contexts/utils';
-import { useOrganizationListInView } from '../../hooks/useOrganizationListInView';
-import { useMosaicEnvironment } from '../hooks/useMosaicEnvironment';
-import { useMosaicRouter } from '../hooks/useMosaicRouter';
-import type {
-  UserButtonBrandingProps,
-  UserButtonCallbacks,
-  UserButtonData,
-  UserButtonInvitation,
-  UserButtonMembership,
-  UserButtonSession,
-  UserButtonSuggestion,
-} from './user-button.types';
+/** The model once Clerk has answered, which is the only shape an action can start from. */
+export type UserButtonReadyModel = Extract<UserButtonModel, { status: 'ready' }>;
 
-// The container awaits these one-shot actions to drive busy state, so the controller exposes their
-// promise; navigation callbacks stay fire-and-forget (`() => void`) and reach the view's DOM handlers.
-interface UserButtonAsyncCallbacks {
-  onSelectOrganization?: (organizationId: string | null) => void | Promise<unknown>;
-  onSwitchSession?: (sessionId: string) => void | Promise<unknown>;
-  onSignOutSession?: (sessionId: string) => void | Promise<unknown>;
-  onSignOutAll?: () => void | Promise<unknown>;
-  onAcceptSuggestion?: (suggestionId: string) => void | Promise<unknown>;
-  onAcceptInvitation?: (invitationId: string) => void | Promise<unknown>;
+export interface UserButtonMachineContext {
+  /** Which action is currently pending. */
+  pendingKey: string | null;
+  /**
+   * The model the action started from. `setActive` swaps the active organization while its
+   * promise is still in flight, so the live model would rearrange the popup mid-action.
+   * The view renders this instead until the action settles.
+   */
+  frozen: UserButtonReadyModel | null;
+  /** Injected per-action effect — the model callback the clicked row runs. */
+  run: () => Promise<unknown>;
+  /** Whether succeeding ends the interaction, and the popup with it. */
+  closeOnSuccess: boolean;
 }
+
+export type UserButtonMachineEvent =
+  | { type: 'OPEN' }
+  | { type: 'CLOSE' }
+  | {
+    type: 'RUN';
+    key: string;
+    frozen: UserButtonReadyModel;
+    run: () => Promise<unknown>;
+    closeOnSuccess: boolean;
+  };
+
+const { createMachine, assign, fromPromise } = setup<UserButtonMachineContext, UserButtonMachineEvent>();
+
+const settled = { pendingKey: null, frozen: null };
+
+export const userButtonMachine = createMachine({
+  id: 'userButton',
+  initial: 'closed',
+  context: {
+    pendingKey: null,
+    frozen: null,
+    run: () => Promise.resolve(),
+    closeOnSuccess: false,
+  },
+  states: {
+    closed: {
+      on: { OPEN: 'open' },
+    },
+    open: {
+      on: {
+        CLOSE: 'closed',
+        RUN: {
+          target: 'busy',
+          actions: assign((_, event) => ({
+            pendingKey: event.key,
+            frozen: event.frozen,
+            run: event.run,
+            closeOnSuccess: event.closeOnSuccess,
+          })),
+        },
+      },
+    },
+    // Reached only from `open`, so a busy popup that is not open is unrepresentable, and RUN going
+    // unhandled here is what stops a second action starting while one is in flight. Dismissing the
+    // popup abandons the action: the request finishes, but nothing is left for its result to land in.
+    busy: {
+      on: { CLOSE: { target: 'closed', actions: assign(() => settled) } },
+      invoke: fromPromise(context => context.run(), {
+        onDone: [
+          { target: 'closed', guard: context => context.closeOnSuccess, actions: assign(() => settled) },
+          { target: 'open', actions: assign(() => settled) },
+        ],
+        // The popup stays up on a failure so the row can be clicked again. Nothing reports what went
+        // wrong yet; the error surface is its own change, and carrying a message before one exists
+        // would mean shipping an untranslated string nobody reads.
+        onError: { target: 'open', actions: assign(() => settled) },
+      }),
+    },
+  },
+});
+
+export type UserButtonControllerOptions = Pick<UserButtonModeProps, 'mode' | 'modePriority'> & UserButtonMenuProps;
 
 export type UserButtonController =
   | { status: 'loading' }
   | { status: 'hidden' }
-  | (UserButtonData &
-      Omit<UserButtonCallbacks, keyof UserButtonAsyncCallbacks> &
-      UserButtonAsyncCallbacks &
-      UserButtonBrandingProps & {
-        status: 'ready';
-        /** Whether the instance has organizations turned on at all. False forces the button to `user` mode. */
-        organizationsEnabled: boolean;
-      });
-
-// Mirrors the `<OrganizationSwitcher>` `afterSelectOrganizationUrl` prop: a full URL/path, a `:token`
-// path template resolved against the organization, or a builder function.
-type AfterSelectUrl<T> = ((entity: T) => string) | string;
+  | ({ status: 'ready' } & Omit<UserButtonViewProps, keyof UserButtonTriggerProps>);
 
 /**
- * How a profile surface opens, in the shape `<UserButton>` and `<OrganizationSwitcher>` already
- * use: a URL is the whole opt-in to navigation, and `modal` forbids one, so the pair can never
- * contradict itself. The two profiles are configured apart, so routing one leaves the other a modal.
- */
-type UserProfileMode =
-  | { userProfileUrl: string; userProfileMode?: 'navigation' }
-  | { userProfileUrl?: never; userProfileMode?: 'modal' };
-
-type OrganizationProfileMode =
-  | { organizationProfileUrl: string; organizationProfileMode?: 'navigation' }
-  | { organizationProfileUrl?: never; organizationProfileMode?: 'modal' };
-
-type CreateOrganizationMode =
-  | { createOrganizationUrl: string; createOrganizationMode?: 'navigation' }
-  | { createOrganizationUrl?: never; createOrganizationMode?: 'modal' };
-
-export type UserButtonControllerOptions = UserProfileMode &
-  OrganizationProfileMode &
-  CreateOrganizationMode & {
-    afterSelectOrganizationUrl?: AfterSelectUrl<OrganizationResource>;
-    /** Where selecting the personal workspace lands. Resolved against the user, not an organization. */
-    afterSelectPersonalUrl?: AfterSelectUrl<UserResource>;
-    /**
-     * Leaves the personal workspace out, for an app whose organizations are the whole product. An
-     * instance that forces organization selection withholds it either way; this cannot opt back in.
-     */
-    hidePersonal?: boolean;
-  };
-
-function resolveAfterSelectUrl<T extends object>(config: AfterSelectUrl<T> | undefined, entity: T): string | undefined {
-  if (typeof config === 'function') {
-    return config(entity);
-  }
-  if (config) {
-    return populateParamFromObject({ urlWithParam: config, entity });
-  }
-  return undefined;
-}
-
-/**
- * One rule for every surface Clerk can host: open the modal unless a URL routes instead. An explicit
- * mode has the last word; a URL on its own means navigation, so passing one is all it takes to
- * route. `url` falls back to Clerk's own so an explicit `navigation` still lands somewhere.
- */
-function openOrNavigate({
-  url,
-  mode,
-  openModal,
-  buildUrl,
-  navigate,
-}: {
-  url: string | undefined;
-  mode: 'navigation' | 'modal' | undefined;
-  openModal: () => void;
-  buildUrl: () => string;
-  navigate: (to: string) => unknown;
-}): () => void {
-  const resolved = mode ?? (url ? 'navigation' : 'modal');
-  return resolved === 'navigation' ? () => void navigate(url ?? buildUrl()) : () => openModal();
-}
-
-const INVITE_MEMBERS_PERMISSION = 'org:sys_memberships:manage';
-
-function displayName(user: UserResource): string {
-  return getFullName(user) || getIdentifier(user);
-}
-
-function toMembership(organization: OrganizationResource): UserButtonMembership {
-  return {
-    kind: 'membership',
-    organizationId: organization.id,
-    name: organization.name,
-    imageUrl: organization.imageUrl || undefined,
-    membersCount: organization.membersCount,
-  };
-}
-
-function toSession(sessionId: string, user: UserResource): UserButtonSession {
-  return {
-    sessionId,
-    name: displayName(user),
-    identifier: getIdentifier(user),
-    imageUrl: user.imageUrl,
-  };
-}
-
-/**
- * @param userProfileCustomPages - The consumer's custom pages, already bridged into clerk-js's
- *   DOM-callback form. The container owns that conversion because it is the layer that can render
- *   the portals behind it, so they arrive here ready to forward and stay out of the public options.
+ * The controller is the layer between the component (view) and the external world (model).
+ * It represents the local state and wraps model actions in order to handle pending states,
+ * keep the UI stable while an action is ongoing, close the popup on completed actions
+ * when appropriate, etc.
  */
 export function useUserButtonController(
-  options?: UserButtonControllerOptions,
-  userProfileCustomPages?: CustomPage[],
+  model: UserButtonModel,
+  options: UserButtonControllerOptions = {},
 ): UserButtonController {
-  const { isLoaded: isUserLoaded, user } = useUser();
-  const { isLoaded: isSessionLoaded, session } = useSession();
-  const { isLoaded: isOrgLoaded, organization } = useOrganization();
-  const { userMemberships, userInvitations, userSuggestions, ref } = useOrganizationListInView();
+  const { mode: requestedMode, modePriority, customMenuItems, menuItemOrder } = options;
+  // The popover's open state and the one action in flight are the same flow: an action that ends the
+  // interaction closes the surface, so they settle together or not at all.
+  const [{ value, context }, send] = useMachine(userButtonMachine);
 
-  const clerk = useClerk();
-  const router = useMosaicRouter();
-  // An app can mount the button inside its own dialog or popover; the modal has to portal into that
-  // same root or it renders behind the surface that opened it.
-  const getContainer = usePortalRoot();
-  const environment = useMosaicEnvironment();
-
-  const manageAccount = openOrNavigate({
-    url: options?.userProfileUrl,
-    mode: options?.userProfileMode,
-    openModal: () => clerk.openUserProfile({ getContainer, customPages: userProfileCustomPages }),
-    buildUrl: () => clerk.buildUserProfileUrl(),
-    navigate: router.navigate,
+  // Every action here is a network round trip, so we can start the
+  // pending state immediately, we use this for the minDuration
+  const displayPendingKey = useSpinDelay(context.pendingKey, {
+    delay: 0,
+    minDuration: context.closeOnSuccess ? 0 : undefined,
   });
 
-  const manageOrganization = openOrNavigate({
-    url: options?.organizationProfileUrl,
-    mode: options?.organizationProfileMode,
-    openModal: () => clerk.openOrganizationProfile({ getContainer }),
-    buildUrl: () => clerk.buildOrganizationProfileUrl(),
-    navigate: router.navigate,
-  });
-
-  const createOrganization = openOrNavigate({
-    url: options?.createOrganizationUrl,
-    mode: options?.createOrganizationMode,
-    openModal: () => clerk.openCreateOrganization({ getContainer }),
-    buildUrl: () => clerk.buildCreateOrganizationUrl(),
-    navigate: router.navigate,
-  });
-
-  // Organizations, single-session and forced selection all come off the environment, and it
-  // hydrates on its own schedule. Waiting for it beats guessing at three answers and rearranging.
-  if (!isUserLoaded || !isSessionLoaded || !isOrgLoaded || !environment) {
-    return { status: 'loading' };
+  if (model.status !== 'ready') {
+    return { status: model.status };
   }
 
-  if (!user || !session) {
-    return { status: 'hidden' };
-  }
+  const close = () => send({ type: 'CLOSE' });
 
-  const { displayConfig, authConfig, organizationSettings } = environment;
-  // clerk-js refuses `setActive({ organization: null })` outright on an instance that forces
-  // organization selection, so there is no personal workspace to offer a way back to.
-  const { enabled: organizationsEnabled, forceOrganizationSelection } = organizationSettings;
-  const { singleSessionMode } = authConfig;
-
-  const canInviteMembers = session.checkAuthorization({ permission: INVITE_MEMBERS_PERMISSION }) ?? false;
-  const membershipData = userMemberships.data ?? [];
-  const suggestionData = userSuggestions.data ?? [];
-  const invitationData = userInvitations.data ?? [];
-
-  const memberships: UserButtonMembership[] = membershipData.map(m => toMembership(m.organization));
-
-  const suggestions: UserButtonSuggestion[] = suggestionData.map(s => ({
-    kind: 'suggestion',
-    id: s.id,
-    organizationId: s.publicOrganizationData.id,
-    name: s.publicOrganizationData.name,
-    imageUrl: s.publicOrganizationData.imageUrl || undefined,
-    status: s.status,
-  }));
-
-  // Accepting is all an invitation row offers, so a revoked or expired one has nothing to offer.
-  const invitations: UserButtonInvitation[] = invitationData.flatMap(i =>
-    i.status === 'pending' || i.status === 'accepted'
-      ? [
-          {
-            kind: 'invitation',
-            id: i.id,
-            status: i.status,
-            organizationId: i.publicOrganizationData.id,
-            organizationName: i.publicOrganizationData.name,
-            imageUrl: i.publicOrganizationData.imageUrl || undefined,
-          },
-        ]
-      : [],
+  const menuItems = customMenuItems?.map(item =>
+    item.href === undefined
+      ? {
+        ...item,
+        onClick: () => {
+          // Always close the menu for custom actions
+          close();
+          item.onClick();
+        },
+      }
+      : item,
   );
 
-  // Organization requests are scoped to the session that makes them, so another account's
-  // workspaces are unknowable until it is the active one. Sessions are all we can hand over.
-  const additionalSessions: UserButtonSession[] = (clerk.client?.signedInSessions ?? []).flatMap(s => {
-    const sessionUser = s.user;
-    if (!sessionUser || s.id === session.id) {
-      return [];
-    }
-    return [toSession(s.id, sessionUser)];
-  });
+  // Wrapper to tie a callback into the machine
+  const runAction = <Args extends unknown[]>(
+    keyFor: (...args: Args) => string,
+    fn: ((...args: Args) => void | Promise<unknown>) | undefined,
+    closeOnSuccess = false,
+  ) =>
+    fn
+      ? (...args: Args) =>
+        send({
+          type: 'RUN',
+          key: keyFor(...args),
+          frozen: model,
+          run: async () => fn(...args),
+          closeOnSuccess,
+        })
+      : undefined;
 
-  // `null` is Clerk's own name for the personal workspace, and it has no organization to resolve
-  // against, so it takes its own URL rather than the organizations'.
-  const afterSelectUrl = (organizationId: string | null): string | undefined => {
-    if (!organizationId) {
-      return resolveAfterSelectUrl(options?.afterSelectPersonalUrl, user);
-    }
-    const selected = membershipData.find(m => m.organization.id === organizationId)?.organization;
-    return selected ? resolveAfterSelectUrl(options?.afterSelectOrganizationUrl, selected) : undefined;
-  };
+  // A callback that wraps a callback so it always closes the popup when done
+  const handOff = (fn: (() => void) | undefined) =>
+    fn
+      ? () => {
+        close();
+        fn();
+      }
+      : undefined;
+
+  // Rendering the model the action froze on holds the popup still while it runs; the result
+  // lands in one step when it settles. See `frozen` in the machine for why.
+  const {
+    status: _status,
+    organizationsEnabled,
+    onSelectOrganization,
+    onSwitchSession,
+    onSignOutSession,
+    onSignOutAll,
+    onAcceptSuggestion,
+    onAcceptInvitation,
+    onManageAccount,
+    onManageOrganization,
+    onInviteMembers,
+    onCreateOrganization,
+    onAddAccount,
+    ...data
+  } = context.frozen ?? model;
+
+  // Force user mode if organizations are disabled
+  const mode = model.organizationsEnabled ? requestedMode : 'user';
 
   return {
     status: 'ready',
-    organizationsEnabled,
-    renderBranding: displayConfig.branded,
-    activeSession: toSession(session.id, user),
-    activeOrganization: organization ? toMembership(organization) : null,
-    // The user resource carries its own memberships, so whether the account has any is settled
-    // before the paginated list is asked. The fetched count still counts, in case the resource is
-    // behind the server.
-    hasOrganizations: user.organizationMemberships.length > 0 || (userMemberships.count ?? 0) > 0,
-    hidePersonal: forceOrganizationSelection || (options?.hidePersonal ?? false),
-    // `isLoading` is "a request is out and nothing has come back", which is the only window where
-    // an empty list is indistinguishable from one that has not arrived. Paging in later pages
-    // leaves it false, since by then the list is already on screen.
-    organizationsLoading: userMemberships.isLoading || userInvitations.isLoading || userSuggestions.isLoading,
-    memberships,
-    suggestions,
-    invitations,
-    additionalSessions,
-    paging: {
-      ref,
-      hasMore: Boolean(userMemberships.hasNextPage || userInvitations.hasNextPage || userSuggestions.hasNextPage),
-    },
-    onSelectOrganization: organizationId =>
-      clerk.setActive({ organization: organizationId, redirectUrl: afterSelectUrl(organizationId) }),
-    onSwitchSession: sessionId =>
-      clerk.setActive({ session: sessionId, redirectUrl: displayConfig.afterSwitchSessionUrl }),
-    onSignOutSession: sessionId =>
-      clerk.signOut({
-        sessionId,
-        // Other accounts stay signed in, so route to the single-session-out URL; otherwise this is
-        // a full sign out.
-        redirectUrl:
-          additionalSessions.length > 0 ? clerk.buildAfterMultiSessionSingleSignOutUrl() : clerk.buildAfterSignOutUrl(),
-      }),
-    // Single-session apps cannot hold a second account, so adding one and signing out of "all
-    // accounts" are meaningless there; the per-account sign out on the active row remains.
-    onSignOutAll: singleSessionMode ? undefined : () => clerk.signOut({ redirectUrl: clerk.buildAfterSignOutUrl() }),
-    onManageAccount: manageAccount,
-    onManageOrganization: manageOrganization,
-    // Invite has no page of its own to route to, so it opens its modal even where managing the
-    // organization is routed to the app's own page.
-    onInviteMembers: canInviteMembers ? () => clerk.openInviteMembers({ getContainer }) : undefined,
-    // The instance can restrict who opens an organization, and the flag also goes false once a user
-    // reaches their creation limit, so it covers both ways the action can be unavailable.
-    onCreateOrganization: user.createOrganizationEnabled ? createOrganization : undefined,
-    onAddAccount: singleSessionMode ? undefined : () => void router.navigate(clerk.buildSignInUrl()),
-    onAcceptSuggestion: suggestionId => {
-      const suggestion = suggestionData.find(s => s.id === suggestionId);
-      return Promise.resolve(suggestion?.accept()).finally(() => void userSuggestions.revalidate?.());
-    },
-    // Accepting an invitation joins the organization, so the membership list is stale too. A
-    // suggestion only files a request an admin has yet to approve, so nothing has been joined.
-    onAcceptInvitation: invitationId => {
-      const invitation = invitationData.find(i => i.id === invitationId);
-      return Promise.resolve(invitation?.accept()).finally(() => {
-        void userInvitations.revalidate?.();
-        void userMemberships.revalidate?.();
-      });
-    },
+    ...data,
+    mode,
+    modePriority,
+    customMenuItems: menuItems,
+    menuItemOrder,
+    open: value !== 'closed',
+    onOpenChange: next => send(next ? { type: 'OPEN' } : { type: 'CLOSE' }),
+    pendingKey: displayPendingKey,
+    onSelectOrganization: runAction(userButtonBusyKeys.selectOrganization, onSelectOrganization, true),
+    onSwitchSession: runAction(userButtonBusyKeys.switchSession, onSwitchSession),
+    onSignOutSession: runAction(userButtonBusyKeys.signOutSession, onSignOutSession),
+    onSignOutAll: runAction(userButtonBusyKeys.signOutAll, onSignOutAll),
+    onAcceptSuggestion: runAction(userButtonBusyKeys.acceptSuggestion, onAcceptSuggestion),
+    onAcceptInvitation: runAction(userButtonBusyKeys.acceptInvitation, onAcceptInvitation),
+    onManageAccount: handOff(onManageAccount),
+    onManageOrganization: handOff(onManageOrganization),
+    onInviteMembers: handOff(onInviteMembers),
+    onCreateOrganization: handOff(onCreateOrganization),
+    onAddAccount: handOff(onAddAccount),
   };
 }
